requirements.txt
Django~=4.0.4
asgiref~=3.5.0
certifi~=2021.10.8
cffi~=1.15.0
coreapi~=2.3.3
coreschema~=0.0.4
cryptography~=37.0.1
defusedxml~=0.7.1
django-cors-headers~=3.11.0
django-filter~=21.1
django-templated-mail~=1.1.1
djangorestframework~=3.13.1
djangorestframework-jsonapi~=5.0.0
djangorestframework-simplejwt~=4.8.0
djoser~=2.1.0
inflection~=0.5.1
itypes~=1.2.0
oauthlib~=3.2.0
pip~=22.0.4
pycparser~=2.21
python3-openid~=3.2.0
pytz~=2022.1
requests~=2.27.1
requests-oauthlib~=1.3.1
setuptools~=57.0.0
six~=1.16.0
social-auth-app-django~=4.0.0
social-auth-core~=4.2.0
sqlparse~=0.4.2
uritemplate~=4.1.1
urllib3~=1.26.9
wheel~=0.36.2